+++ src/endstone_python/endstone_python.cpp
@@ -35,7 +35,7 @@ void init_game_mode(py::module_ &);
 void init_logger(py::module_ &);
 
 void init_util(py::module_ &);
-void init_command(py::module_ &);
+void init_command(py::module &, py::class_<CommandSender, Permissible> &command_sender);
 void init_plugin(py::module_ &);
 void init_scheduler(py::module_ &);
 void init_permissions(py::module_ &, py::class_<Permissible> &permissible, py::class_<Permission> &permission,
@@ -54,23 +54,25 @@ PYBIND11_MODULE(endstone_python, m)  // NOLINT(*-use-anonymous-namespace)
     auto event_priority = py::enum_<EventPriority>(
         m, "EventPriority",
         "Listeners are called in following order: LOWEST -> LOW -> NORMAL -> HIGH -> HIGHEST -> MONITOR");
-    auto server = py::class_<Server>(m, "Server");
+
     auto permissible = py::class_<Permissible>(m, "Permissible");
+    auto command_sender = py::class_<CommandSender, Permissible>(m, "CommandSender");
     auto permission = py::class_<Permission>(m, "Permission");
     auto permission_default = py::enum_<PermissionDefault>(m, "PermissionDefault");
+    auto server = py::class_<Server>(m, "Server");
 
     init_color_format(m);
     init_game_mode(m);
     init_logger(m);
 
     init_util(m);
-    init_command(m);
 
+    init_actor(m);
+    init_player(m);
+    init_command(m, command_sender);
     init_plugin(m);
     init_scheduler(m);
     init_permissions(m, permissible, permission, permission_default);
-    init_actor(m);
-    init_player(m);
     init_level(m);
     init_server(server);
     init_event(m, event, event_priority);
@@ -203,7 +205,8 @@ void init_player(py::module_ &m)
         .def_property_readonly("address", &Player::getAddress, "Gets the socket address of this player")
         .def("send_raw_message", &Player::sendRawMessage, py::arg("message"), "Sends this player a raw message")
         .def("send_popup", &Player::sendPopup, py::arg("message"), "Sends this player a popup message")
-        .def("send_tip", &Player::sendTip, py::arg("message"), "Sends this player a tip message");
+        .def("send_tip", &Player::sendTip, py::arg("message"), "Sends this player a tip message")
+        .def("kick", &Player::kick, py::arg("message"), "Kicks player with custom kick message.");
 }
 
 }  // namespace endstone::detail