+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - `Server::getOnlinePlayers` for getting all online players.
+- `Player::kick` for kicking a specific player with a custom message.
 
 ### Changed